version for release 28.0.0

+++ Rakefile
@@ -27,8 +27,8 @@ namespace :integration do
   desc 'Run Test Kitchen with Vagrant'
   task :vagrant do
     options = {
-      versions: %w(16 17 18 19 20 21 22 23 24 25 26),
-      platforms: %w(centos-69 centos-7),
+      versions: %w(16 17 18 19 20 21 22 23 24 25 26 27 28),
+      platforms: %w(centos-7),
     }
     opts = OptionParser.new
     opts.banner = 'Usage: rake integration:vagrant [options]'
@@ -91,7 +91,7 @@ namespace :integration do
           old_instance = instance
         end
         # to improve the odds it actually destroys properly, rename it back to original
-        if File.exist?(".kitchen/#{old_instance.name}.yml")
+        if !old_instance.nil? && File.exist?(".kitchen/#{old_instance.name}.yml")
           puts "Done with #{ver} on #{plat}. Moving #{old_instance.name}.yml to #{first_instance.name}.yml before destroying."
           File.rename(".kitchen/#{old_instance.name}.yml", ".kitchen/#{first_instance.name}.yml")
           first_instance.destroy
